backend/requirements.txt
Django==3.2.4
djangorestframework==3.13.1
django-cors-headers==3.10.1
#django-mysql==3.3.0
django-phonenumber-field==4.0.0
django-jsonfield==1.4.1
django-phonenumbers
#mysqlclient==2.0.1
psycopg2-binary
requests==2.27.1

jsonfield2==3.1.0

django-ipware==4.0.2

django-filter==21.1 

drf-yasg==1.20.0 
ptvsd==4.3.2

transitions==0.8.10

validators==0.19.0 

PyJWT==2.0.1
redis~=4.2.2
python-dateutil==2.8.2

gunicorn==20.0.4

pandas==1.1.3

python-dotenv==1.0.1

djangorestframework-simplejwt==5.3.1